requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
amqp==5.2.0
    # via kombu
appdirs==1.4.4
    # via fs
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via openedx-events
backports-zoneinfo[tzdata]==0.2.1 ; python_version < "3.9"
    # via
    #   -c requirements/constraints.txt
    #   celery
    #   django
    #   djangorestframework
    #   kombu
billiard==4.2.0
    # via celery
boto3==1.34.135
    # via fs-s3fs
botocore==1.34.135
    # via
    #   boto3
    #   s3transfer
celery==5.4.0
    # via
    #   -r requirements/base.in
    #   event-tracking
certifi==2024.6.2
    # via requests
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   code-annotations
    #   edx-django-utils
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
code-annotations==1.8.0
    # via edx-toggles
cryptography==42.0.8
    # via pyjwt
django==4.2.13
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.in
    #   django-crum
    #   django-model-utils
    #   django-waffle
    #   djangorestframework
    #   drf-jwt
    #   edx-completion
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-toggles
    #   event-tracking
    #   openedx-django-pyfs
    #   openedx-events
    #   openedx-filters
django-crum==0.7.9
    # via
    #   edx-django-utils
    #   edx-toggles
django-model-utils==4.5.1
    # via edx-completion
django-waffle==4.1.0
    # via
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-toggles
djangorestframework==3.15.2
    # via
    #   -r requirements/base.in
    #   drf-jwt
    #   edx-completion
    #   edx-drf-extensions
drf-jwt==1.19.2
    # via edx-drf-extensions
edx-ccx-keys==1.3.0
    # via openedx-events
edx-completion==4.6.5
    # via -r requirements/base.in
edx-django-utils==5.14.2
    # via
    #   edx-drf-extensions
    #   edx-toggles
    #   event-tracking
    #   openedx-events
edx-drf-extensions==10.3.0
    # via
    #   -r requirements/base.in
    #   edx-completion
edx-opaque-keys[django]==2.10.0
    # via
    #   edx-ccx-keys
    #   edx-completion
    #   edx-drf-extensions
    #   openedx-events
edx-toggles==5.2.0
    # via
    #   edx-completion
    #   event-tracking
event-tracking==2.4.0
    # via edx-completion
fastavro==1.9.4
    # via openedx-events
fs==2.4.16
    # via
    #   fs-s3fs
    #   openedx-django-pyfs
    #   xblock
fs-s3fs==1.1.1
    # via openedx-django-pyfs
idna==3.7
    # via requests
jinja2==3.1.4
    # via code-annotations
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.3.7
    # via celery
lazy==1.6
    # via xblock
lxml==5.2.2
    # via xblock
mako==1.3.5
    # via
    #   xblock
    #   xblock-utils
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
    #   xblock
newrelic==9.11.0
    # via edx-django-utils
openedx-django-pyfs==3.6.0
    # via xblock
openedx-events==9.11.0
    # via event-tracking
openedx-filters==1.9.0
    # via -r requirements/base.in
pbr==6.0.0
    # via stevedore
prompt-toolkit==3.0.47
    # via click-repl
psutil==6.0.0
    # via edx-django-utils
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.8.0
    # via
    #   drf-jwt
    #   edx-drf-extensions
pymongo==3.13.0
    # via
    #   edx-opaque-keys
    #   event-tracking
pynacl==1.5.0
    # via edx-django-utils
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
    #   xblock
python-slugify==8.0.4
    # via code-annotations
pytz==2024.1
    # via
    #   edx-completion
    #   event-tracking
    #   xblock
pyyaml==6.0.1
    # via
    #   code-annotations
    #   xblock
requests==2.32.3
    # via edx-drf-extensions
s3transfer==0.10.2
    # via boto3
semantic-version==2.10.0
    # via edx-drf-extensions
simplejson==3.19.2
    # via
    #   xblock
    #   xblock-utils
six==1.16.0
    # via
    #   edx-ccx-keys
    #   event-tracking
    #   fs
    #   fs-s3fs
    #   python-dateutil
sqlparse==0.5.0
    # via django
stevedore==5.2.0
    # via
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.12.2
    # via
    #   asgiref
    #   edx-opaque-keys
    #   kombu
tzdata==2024.1
    # via
    #   backports-zoneinfo
    #   celery
urllib3==1.26.19
    # via
    #   botocore
    #   requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
web-fragments==2.2.0
    # via
    #   xblock
    #   xblock-utils
webob==1.8.7
    # via xblock
xblock[django]==4.0.1
    # via
    #   -r requirements/base.in
    #   edx-completion
    #   xblock-utils
xblock-utils==4.0.0
    # via -r requirements/base.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
